Second Commit, authorization by permission types, Requests, Events and Listeners added.

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Events\ShowMessageEvent;
+use App\Http\Requests\StorePostRequest;
 use App\Models\Post;
 use Illuminate\Http\Request;
 
@@ -31,28 +32,35 @@ class PostController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->cannot('create', Post::class)) {
+            return redirect()->route('profile');
+        }
+
         return view('posts.addPost');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StorePostRequest $request)
     {
 
-        $this->validate($request, [
-            'title' => ['required'],
-            'description' => ['required'],
-        ]);
+        //$this->validate($request, [
+        //    'title' => ['required'],
+        //    'description' => ['required'],
+        //]);
+        //$validiated = $request->validate();
+
+        //$this->authorize('create', Post::class);
+
+        $validiated = $request->validated();
         $post = $request->only('title', 'description');
         $post['public'] = 1;
         $post['user_id'] = auth()->user()->id;
         Post::create($post);
-
-//        event(new ShowMessageEvent($post));
 
         return redirect()->route('profile');
     }
@@ -60,12 +68,13 @@ class PostController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
         $post = Post::with('user')->where('id', $id)->get();
+
         // dd($post);
         return $post;
     }
@@ -73,13 +82,18 @@ class PostController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
+        if (auth()->user()->cannot('edit', Post::class)) {
+
+            return redirect()->route('profile');
+        }
+
         $post = Post::with('user')->find($id);
-        // dd($user['description']);
+
         return view('posts.updatePost', [
             'post' => $post,
         ]);
@@ -88,8 +102,8 @@ class PostController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -104,7 +118,7 @@ class PostController extends Controller
 
 //        dd(Post::find($id));
 
-        event(new ShowMessageEvent($id,'edit'));
+        event(new ShowMessageEvent($id, 'edit'));
 
         return redirect()->route("profile");
     }
@@ -112,12 +126,17 @@ class PostController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        event(new ShowMessageEvent($id,'delete'));
+        if (auth()->user()->cannot('delete', Post::class)) {
+
+            return redirect()->route('profile');
+        }
+
+        event(new ShowMessageEvent($id, 'delete'));
 
         Post::find($id)->delete();
